Moved around files. Added skeleton files for sign up controller, presenter and view model.
