+++ src/main/java/interface_adapter/controller/UserSignupController.java
@@ -1,0 +1,4 @@
+package interface_adapter.controller;
+
+public class UserSignupController {
+}
