+++ src/main/java/interface_adapter/presenter/UserSignupPresenter.java
@@ -1,0 +1,4 @@
+package interface_adapter.presenter;
+
+public class UserSignupPresenter {
+}
