+++ src/main/java/interface_adapter/viewmodel/UserSignupViewModel.java
@@ -1,0 +1,4 @@
+package interface_adapter.viewmodel;
+
+public class UserSignupViewModel {
+}
